Unify domain logic, extract registry service, pass S3 key to Modal
Three architectural improvements:

1. Shared dhub-core package: Extract duplicated manifest parsing and
   domain models (SkillManifest, RuntimeConfig, etc.) into a shared
   workspace package. Client and server re-export from dhub_core,
   eliminating drift between the two independent copies.

2. Registry service extraction: Move business logic helpers out of
   registry_routes.py (1022 lines) into registry_service.py. Route
   handlers become thin adapters that parse requests and delegate.
   Gauntlet pipeline, org membership checks, quarantine logic, and
   eval triggering are now testable without HTTP mocking.

3. S3 key instead of raw bytes for Modal eval: The publish endpoint
   no longer serializes the entire zip (up to 50 MB) through Modal's
   parameter transport. Instead, it passes the S3 key and bucket,
   and the eval container downloads directly from S3.
