+++ server/pyproject.toml
@@ -17,6 +17,7 @@ dependencies = [
     "cryptography>=42.0.0",
     "pyyaml>=6.0.0",
     "modal>=1.3.2",
+    "dhub-core",
 ]
 
 [project.optional-dependencies]
@@ -34,6 +35,9 @@ build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
 packages = ["src/decision_hub"]
+
+[tool.uv.sources]
+dhub-core = { workspace = true }
 
 [tool.pytest.ini_options]
 testpaths = ["tests"]
