+++ server/src/decision_hub/api/registry_service.py
@@ -1,0 +1,486 @@
+"""Registry business logic — extracted from registry_routes.py.
+
+Pure functions that handle publishing, validation, gauntlet pipeline,
+and eval triggering. Route handlers call these instead of inlining
+the logic, making the business rules testable without HTTP mocking.
+"""
+
+import logging
+import tempfile
+from pathlib import Path
+from uuid import UUID
+
+from fastapi import HTTPException
+from sqlalchemy.engine import Connection
+
+from decision_hub.domain.gauntlet import run_static_checks
+from decision_hub.domain.publish import build_quarantine_s3_key
+from decision_hub.domain.skill_manifest import extract_body, extract_description, parse_skill_md
+from decision_hub.infra.database import (
+    find_org_by_slug,
+    find_org_member,
+    insert_audit_log,
+)
+from decision_hub.infra.storage import upload_skill_zip
+from decision_hub.models import GauntletReport, Organization
+from decision_hub.settings import Settings
+
+logger = logging.getLogger(__name__)
+
+
+def require_org_membership(
+    conn: Connection,
+    org_slug: str,
+    user_id: UUID,
+    *,
+    admin_only: bool = False,
+) -> Organization:
+    """Verify org exists and user is a member; return the Organisation.
+
+    Raises 404 if org not found, 403 if not a member (or not admin
+    when admin_only=True).
+    """
+    org = find_org_by_slug(conn, org_slug)
+    if org is None:
+        raise HTTPException(status_code=404, detail="Organisation not found")
+
+    member = find_org_member(conn, org.id, user_id)
+    if member is None:
+        raise HTTPException(
+            status_code=403,
+            detail="You are not a member of this organisation",
+        )
+    if admin_only and member.role not in ("owner", "admin"):
+        raise HTTPException(
+            status_code=403,
+            detail="Only org owners and admins can perform this action",
+        )
+    return org
+
+
+def parse_manifest_from_content(
+    skill_md_content: str,
+    file_bytes: bytes,
+) -> tuple[dict | None, object | None, tuple]:
+    """Parse SKILL.md and extract runtime config, eval config, and eval cases.
+
+    Uses a temp file because parse_skill_md expects a file path.
+    Returns (runtime_config_dict, eval_config, eval_cases).
+
+    Raises HTTPException(422) if the manifest is malformed — fail-closed
+    to prevent publishing skills with unparseable manifests.
+    """
+    with tempfile.NamedTemporaryFile(mode="w", suffix=".md", delete=False) as tmp:
+        tmp.write(skill_md_content)
+        tmp_path = Path(tmp.name)
+
+    try:
+        manifest = parse_skill_md(tmp_path)
+        return (
+            extract_runtime_config_dict(manifest),
+            extract_assessment_config(manifest),
+            try_parse_assessment_cases(file_bytes),
+        )
+    except ValueError as exc:
+        logger.warning("Manifest parse failed (rejecting publish): %s", exc)
+        raise HTTPException(
+            status_code=422,
+            detail=f"SKILL.md manifest is malformed: {exc}",
+        )
+    finally:
+        tmp_path.unlink()
+
+
+def run_gauntlet_pipeline(
+    skill_md_content: str,
+    lockfile_content: str | None,
+    source_files: list[tuple[str, str]],
+    skill_name: str,
+    description: str,
+    skill_md_body: str,
+    settings: Settings,
+) -> tuple[GauntletReport, list[dict], dict | None]:
+    """Run Gauntlet static checks and serialize results for audit logging.
+
+    Returns (report, check_results_dicts, llm_reasoning).
+    """
+    report = run_static_checks(
+        skill_md_content,
+        lockfile_content,
+        source_files,
+        skill_name=skill_name,
+        skill_description=description,
+        analyze_fn=_build_analyze_fn(settings),
+        skill_md_body=skill_md_body,
+        allowed_tools=None,
+        analyze_prompt_fn=_build_analyze_prompt_fn(settings),
+        is_verified_org=True,
+    )
+
+    check_results_dicts = [
+        {
+            "check_name": r.check_name,
+            "severity": r.severity,
+            "message": r.message,
+        }
+        for r in report.results
+    ]
+
+    llm_reasoning = {
+        r.check_name: r.details
+        for r in report.results
+        if r.details is not None
+    } or None
+
+    return report, check_results_dicts, llm_reasoning
+
+
+def quarantine_rejected_skill(
+    conn: Connection,
+    s3_client,
+    bucket: str,
+    file_bytes: bytes,
+    *,
+    org_slug: str,
+    skill_name: str,
+    version: str,
+    report: GauntletReport,
+    check_results: list[dict],
+    llm_reasoning: dict | None,
+    publisher: str,
+) -> None:
+    """Upload rejected zip to quarantine, log the rejection, and raise 422.
+
+    Inserts and commits the audit log before uploading to quarantine S3,
+    so the rejection record is durable even if the S3 upload fails.
+    """
+    q_key = build_quarantine_s3_key(org_slug, skill_name, version)
+
+    insert_audit_log(
+        conn,
+        org_slug=org_slug,
+        skill_name=skill_name,
+        semver=version,
+        grade=report.grade,
+        check_results=check_results,
+        publisher=publisher,
+        version_id=None,
+        llm_reasoning=llm_reasoning,
+        quarantine_s3_key=q_key,
+    )
+    # Commit the audit record before raising (or uploading to S3) so it
+    # survives the transaction rollback that engine.begin() performs on
+    # exception. This ensures rejection forensics are always preserved.
+    conn.commit()
+
+    upload_skill_zip(s3_client, bucket, q_key, file_bytes)
+
+    raise HTTPException(
+        status_code=422,
+        detail=f"Gauntlet checks failed: {report.summary}",
+    )
+
+
+def _build_analyze_fn(settings: Settings):
+    """Build a Gemini analyze callback if google_api_key is configured.
+
+    Returns None if no API key is set, which causes the safety scan
+    to run in strict regex-only mode.
+    """
+    if not settings.google_api_key:
+        return None
+
+    from decision_hub.infra.gemini import analyze_code_safety, create_gemini_client
+
+    gemini_client = create_gemini_client(settings.google_api_key)
+
+    def analyze_fn(snippets, skill_name, skill_description):
+        return analyze_code_safety(
+            gemini_client,
+            snippets,
+            skill_name,
+            skill_description,
+            model=settings.gemini_model,
+        )
+
+    return analyze_fn
+
+
+def _build_analyze_prompt_fn(settings: Settings):
+    """Build a Gemini prompt analyze callback if google_api_key is configured.
+
+    Returns None if no API key is set, which causes the prompt safety scan
+    to run in strict regex-only mode.
+    """
+    if not settings.google_api_key:
+        return None
+
+    from decision_hub.infra.gemini import analyze_prompt_safety, create_gemini_client
+
+    gemini_client = create_gemini_client(settings.google_api_key)
+
+    def analyze_prompt_fn(prompt_hits, skill_name, skill_description):
+        return analyze_prompt_safety(
+            gemini_client,
+            prompt_hits,
+            skill_name,
+            skill_description,
+            model=settings.gemini_model,
+        )
+
+    return analyze_prompt_fn
+
+
+def extract_runtime_config_dict(manifest) -> dict | None:
+    """Extract runtime config as a JSON-compatible dict for database storage."""
+    if manifest.runtime is None:
+        return None
+
+    runtime_dict = {
+        "language": manifest.runtime.language,
+        "entrypoint": manifest.runtime.entrypoint,
+        "version_hint": manifest.runtime.version_hint,
+        "env": list(manifest.runtime.env),
+        "capabilities": list(manifest.runtime.capabilities),
+        "repair_strategy": manifest.runtime.repair_strategy,
+    }
+
+    if manifest.runtime.dependencies:
+        runtime_dict["dependencies"] = {
+            "system": list(manifest.runtime.dependencies.system),
+            "package_manager": manifest.runtime.dependencies.package_manager,
+            "packages": list(manifest.runtime.dependencies.packages),
+            "lockfile": manifest.runtime.dependencies.lockfile,
+        }
+
+    return runtime_dict
+
+
+def extract_assessment_config(manifest):
+    """Extract eval config from manifest (returns None if not present)."""
+    return manifest.evals
+
+
+def try_parse_assessment_cases(file_bytes: bytes):
+    """Parse eval cases from zip. Returns empty tuple if no evals/ directory.
+
+    Raises HTTPException(422) if eval files exist but are malformed —
+    fail-closed to prevent bypassing the eval pipeline with broken YAML.
+    """
+    from decision_hub.domain.skill_manifest import parse_eval_cases_from_zip
+
+    try:
+        return parse_eval_cases_from_zip(file_bytes)
+    except ValueError as exc:
+        logger.warning("Eval case parse failed (rejecting publish): %s", exc)
+        raise HTTPException(
+            status_code=422,
+            detail=f"Eval case files are malformed: {exc}",
+        )
+
+
+def maybe_trigger_agent_assessment(
+    eval_config,
+    eval_cases: tuple,
+    s3_key: str,
+    s3_bucket: str,
+    version_id,
+    org_slug: str,
+    skill_name: str,
+    settings: Settings,
+    user_id,
+):
+    """Conditionally trigger background agent evaluation if eval config present.
+
+    Uses Modal's ``Function.spawn()`` so the eval runs in its own container,
+    fully independent of the web server's lifecycle.  The caller must commit
+    the version row before calling this function.
+
+    Passes the S3 key (not the raw zip bytes) to the Modal function, so
+    the 50 MB payload is fetched from S3 inside the eval container instead
+    of being serialized through Modal's parameter transport.
+
+    Raises HTTPException(422) if eval config is declared in the manifest
+    but no valid eval case files were found — prevents bypassing the eval
+    pipeline by omitting case files while keeping the config.
+    """
+    if eval_config and not eval_cases:
+        raise HTTPException(
+            status_code=422,
+            detail="Eval config declared in manifest but no eval case files found in evals/",
+        )
+    if eval_config and eval_cases:
+        import modal
+
+        # Serialize EvalCase dataclasses to dicts for Modal transport
+        cases_dicts = [
+            {
+                "name": c.name,
+                "description": c.description,
+                "prompt": c.prompt,
+                "judge_criteria": c.judge_criteria,
+            }
+            for c in eval_cases
+        ]
+
+        run_eval = modal.Function.from_name(
+            settings.modal_app_name,
+            "run_eval_task",
+        )
+        run_eval.spawn(
+            version_id=str(version_id),
+            eval_agent=eval_config.agent,
+            eval_judge_model=eval_config.judge_model,
+            eval_cases_dicts=cases_dicts,
+            s3_key=s3_key,
+            s3_bucket=s3_bucket,
+            org_slug=org_slug,
+            skill_name=skill_name,
+            user_id=str(user_id),
+        )
+        return "pending"
+    return None
+
+
+def run_assessment_background(
+    version_id,
+    eval_config,
+    eval_cases: tuple,
+    skill_zip: bytes,
+    org_slug: str,
+    skill_name: str,
+    settings: Settings,
+    user_id,
+):
+    """Background task to run agent assessments and store report."""
+    from cryptography.fernet import Fernet
+
+    from decision_hub.domain.evals import run_eval_pipeline
+    from decision_hub.infra.database import create_engine, get_api_keys_for_eval
+    from decision_hub.infra.modal_client import get_agent_config, validate_api_key
+
+    try:
+        print(f"[assessment] Phase 1: loading API keys for {org_slug}/{skill_name}", flush=True)
+        engine = create_engine(settings.database_url)
+
+        # --- Phase 1: read API keys then release the connection ---
+        # The connection must be closed before the pipeline runs because
+        # Modal sandbox + agent execution takes 5-10 minutes. PgBouncer
+        # kills idle-in-transaction connections well before that.
+        agent_config = get_agent_config(eval_config.agent)
+        required_keys = [agent_config.key_env_var] if agent_config.key_env_var else []
+        # The judge always calls the Anthropic API, so we need an
+        # ANTHROPIC_API_KEY even when the agent under test uses a different
+        # provider (e.g. codex uses CODEX_API_KEY, gemini uses GEMINI_API_KEY).
+        judge_key_name = "ANTHROPIC_API_KEY"
+        if judge_key_name not in required_keys:
+            required_keys.append(judge_key_name)
+        with engine.connect() as conn:
+            encrypted_keys = get_api_keys_for_eval(conn, user_id, required_keys)
+            conn.commit()
+
+        print(f"[assessment] Got {len(encrypted_keys)} keys: {list(encrypted_keys.keys())}", flush=True)
+
+        fernet = Fernet(settings.fernet_key.encode())
+        agent_env_vars = {
+            name: fernet.decrypt(value).decode()
+            for name, value in encrypted_keys.items()
+        }
+
+        # Fail fast if the API key is invalid instead of hanging in a sandbox
+        for key_name, key_value in agent_env_vars.items():
+            validate_api_key(key_name, key_value)
+        print("[assessment] API key validation passed", flush=True)
+
+        # --- Phase 2: run pipeline (no DB connection held) ---
+        print(f"[assessment] Phase 2: running pipeline ({len(eval_cases)} cases)", flush=True)
+        judge_api_key = agent_env_vars.get(judge_key_name, "")
+        case_results, passed, total, total_duration_ms = run_eval_pipeline(
+            skill_zip=skill_zip,
+            eval_config=eval_config,
+            eval_cases=eval_cases,
+            agent_env_vars=agent_env_vars,
+            org_slug=org_slug,
+            skill_name=skill_name,
+            judge_api_key=judge_api_key,
+        )
+
+        # --- Phase 3: store results in a fresh connection ---
+        all_passed = all(r["verdict"] == "pass" for r in case_results)
+        status = "completed" if all_passed else "failed"
+
+        print(f"[assessment] Phase 3: storing results — {passed}/{total} passed, status={status}", flush=True)
+        with engine.connect() as conn:
+            from decision_hub.infra.database import insert_eval_report
+
+            insert_eval_report(
+                conn,
+                version_id=version_id,
+                agent=eval_config.agent,
+                judge_model=eval_config.judge_model,
+                case_results=case_results,
+                passed=passed,
+                total=total,
+                total_duration_ms=total_duration_ms,
+                status=status,
+            )
+            conn.commit()
+
+        print(f"[assessment] Done — {passed}/{total} passed in {total_duration_ms}ms", flush=True)
+
+    except Exception as e:
+        logger.error(f"Agent assessment failed for version {version_id}: {e}")
+        print(f"[assessment] ERROR: {e}", flush=True)
+        # INSERT an error report — no row exists yet because the failure
+        # happened before insert_eval_report() was reached in the happy path.
+        try:
+            from decision_hub.infra.database import create_engine as _create_engine
+            from decision_hub.infra.database import insert_eval_report
+
+            err_engine = _create_engine(settings.database_url)
+            with err_engine.connect() as err_conn:
+                insert_eval_report(
+                    err_conn,
+                    version_id=version_id,
+                    agent=eval_config.agent,
+                    judge_model=eval_config.judge_model,
+                    case_results=[],
+                    passed=0,
+                    total=len(eval_cases),
+                    total_duration_ms=0,
+                    status="failed",
+                    error_message=str(e),
+                )
+                err_conn.commit()
+        except Exception as inner:
+            logger.error(
+                f"Failed to store error report for version {version_id}: {inner}"
+            )
+
+
+def reconstruct_runtime_config(runtime_dict: dict | None):
+    """Reconstruct a RuntimeConfig from a stored dict (for testing/validation)."""
+    if runtime_dict is None:
+        return None
+
+    from decision_hub.models import DependencySpec, RuntimeConfig
+
+    deps_dict = runtime_dict.get("dependencies")
+    dependencies = None
+    if deps_dict:
+        dependencies = DependencySpec(
+            system=tuple(deps_dict.get("system", [])),
+            package_manager=deps_dict.get("package_manager", ""),
+            packages=tuple(deps_dict.get("packages", [])),
+            lockfile=deps_dict.get("lockfile"),
+        )
+
+    return RuntimeConfig(
+        language=runtime_dict.get("language", ""),
+        entrypoint=runtime_dict.get("entrypoint", ""),
+        version_hint=runtime_dict.get("version_hint"),
+        env=tuple(runtime_dict.get("env", [])),
+        capabilities=tuple(runtime_dict.get("capabilities", [])),
+        dependencies=dependencies,
+        repair_strategy=runtime_dict.get("repair_strategy", "attempt_install"),
+    )
