+++ server/src/decision_hub/models.py
@@ -5,6 +5,16 @@ from datetime import datetime
 from typing import Literal
 from uuid import UUID
 
+# Re-export shared models from dhub_core (single source of truth)
+from dhub_core.models import (  # noqa: F401
+    AgentTestTarget,
+    DependencySpec,
+    EvalConfig,
+    RuntimeConfig,
+    SkillManifest,
+    TestingConfig,
+)
+
 # Status vocabularies — single source of truth for magic strings
 CheckSeverity = Literal["pass", "warn", "fail"]
 SafetyGrade = Literal["A", "B", "C", "F"]
@@ -89,59 +99,6 @@ class AgentSandboxConfig:
 
 
 @dataclass(frozen=True)
-class SkillManifest:
-    """Parsed SKILL.md content."""
-    name: str
-    description: str
-    license: str | None
-    compatibility: str | None
-    metadata: dict[str, str] | None
-    allowed_tools: str | None
-    runtime: "RuntimeConfig | None"
-    evals: "EvalConfig | None"
-    body: str
-    testing: "TestingConfig | None" = None  # Legacy field for backward compatibility
-
-
-# ---------------------------------------------------------------------------
-# Soft contract runtime
-# ---------------------------------------------------------------------------
-
-
-@dataclass(frozen=True)
-class DependencySpec:
-    """Declared dependencies for a skill's runtime environment."""
-    system: tuple[str, ...]
-    package_manager: str
-    packages: tuple[str, ...]
-    lockfile: str | None
-
-
-@dataclass(frozen=True)
-class RuntimeConfig:
-    """Soft contract: what the skill needs to run."""
-    language: str
-    entrypoint: str
-    version_hint: str | None = None
-    env: tuple[str, ...] = ()
-    capabilities: tuple[str, ...] = ()
-    dependencies: DependencySpec | None = None
-    repair_strategy: str = "attempt_install"
-
-
-# ---------------------------------------------------------------------------
-# Agent evals
-# ---------------------------------------------------------------------------
-
-
-@dataclass(frozen=True)
-class EvalConfig:
-    """Eval configuration from SKILL.md frontmatter."""
-    agent: str
-    judge_model: str
-
-
-@dataclass(frozen=True)
 class EvalCase:
     """A single evaluation case from evals/*.yaml."""
     name: str
@@ -178,26 +135,6 @@ class EvalReport:
     status: EvalReportStatus
     error_message: str | None = None
     created_at: datetime | None = None
-
-
-# ---------------------------------------------------------------------------
-# Legacy models for backward compatibility
-# ---------------------------------------------------------------------------
-
-
-@dataclass(frozen=True)
-class AgentTestTarget:
-    """Legacy testing target configuration."""
-    name: str
-    required_keys: tuple[str, ...]
-
-
-@dataclass(frozen=True)
-class TestingConfig:
-    """Legacy testing configuration from SKILL.md frontmatter."""
-    __test__ = False  # prevent pytest from trying to collect this dataclass
-    cases: str
-    agents: tuple[AgentTestTarget, ...]
 
 
 @dataclass(frozen=True)
