+++ server/tests/test_api/test_registry_routes.py
@@ -111,8 +111,8 @@ class TestPublishSkill:
         assert resp.status_code == 503
         assert "LLM judge" in resp.json()["detail"]
 
-    @patch("decision_hub.api.registry_routes._build_analyze_prompt_fn", return_value=None)
-    @patch("decision_hub.api.registry_routes._build_analyze_fn", return_value=None)
+    @patch("decision_hub.api.registry_service._build_analyze_prompt_fn", return_value=None)
+    @patch("decision_hub.api.registry_service._build_analyze_fn", return_value=None)
     @patch("decision_hub.api.registry_routes.insert_audit_log")
     @patch("decision_hub.api.registry_routes.update_skill_description")
     @patch("decision_hub.api.registry_routes.insert_version")
@@ -120,8 +120,8 @@ class TestPublishSkill:
     @patch("decision_hub.api.registry_routes.find_skill")
     @patch("decision_hub.api.registry_routes.upload_skill_zip")
     @patch("decision_hub.api.registry_routes.compute_checksum")
-    @patch("decision_hub.api.registry_routes.find_org_member")
-    @patch("decision_hub.api.registry_routes.find_org_by_slug")
+    @patch("decision_hub.api.registry_service.find_org_member")
+    @patch("decision_hub.api.registry_service.find_org_by_slug")
     def test_publish_success(
         self,
         mock_find_org: MagicMock,
@@ -166,7 +166,7 @@ class TestPublishSkill:
         # Audit log inserted for successful publish
         mock_insert_audit.assert_called_once()
 
-    @patch("decision_hub.api.registry_routes.find_org_by_slug")
+    @patch("decision_hub.api.registry_service.find_org_by_slug")
     def test_publish_org_not_found(
         self,
         mock_find_org: MagicMock,
@@ -183,8 +183,8 @@ class TestPublishSkill:
         assert resp.status_code == 404
         assert "Organisation not found" in resp.json()["detail"]
 
-    @patch("decision_hub.api.registry_routes.find_org_member")
-    @patch("decision_hub.api.registry_routes.find_org_by_slug")
+    @patch("decision_hub.api.registry_service.find_org_member")
+    @patch("decision_hub.api.registry_service.find_org_by_slug")
     def test_publish_not_member(
         self,
         mock_find_org: MagicMock,
@@ -220,8 +220,8 @@ class TestPublishSkill:
 
         assert resp.status_code == 401
 
-    @patch("decision_hub.api.registry_routes.find_org_member")
-    @patch("decision_hub.api.registry_routes.find_org_by_slug")
+    @patch("decision_hub.api.registry_service.find_org_member")
+    @patch("decision_hub.api.registry_service.find_org_by_slug")
     def test_publish_rejects_oversized_upload(
         self,
         mock_find_org: MagicMock,
@@ -245,8 +245,8 @@ class TestPublishSkill:
         assert resp.status_code == 413
         assert "maximum size" in resp.json()["detail"]
 
-    @patch("decision_hub.api.registry_routes._build_analyze_prompt_fn", return_value=None)
-    @patch("decision_hub.api.registry_routes._build_analyze_fn", return_value=None)
+    @patch("decision_hub.api.registry_service._build_analyze_prompt_fn", return_value=None)
+    @patch("decision_hub.api.registry_service._build_analyze_fn", return_value=None)
     @patch("decision_hub.api.registry_routes.insert_audit_log")
     @patch("decision_hub.api.registry_routes.insert_version")
     @patch("decision_hub.api.registry_routes.find_version")
@@ -254,8 +254,8 @@ class TestPublishSkill:
     @patch("decision_hub.api.registry_routes.find_skill")
     @patch("decision_hub.api.registry_routes.upload_skill_zip")
     @patch("decision_hub.api.registry_routes.compute_checksum")
-    @patch("decision_hub.api.registry_routes.find_org_member")
-    @patch("decision_hub.api.registry_routes.find_org_by_slug")
+    @patch("decision_hub.api.registry_service.find_org_member")
+    @patch("decision_hub.api.registry_service.find_org_by_slug")
     def test_publish_creates_new_skill(
         self,
         mock_find_org: MagicMock,
@@ -301,13 +301,13 @@ class TestPublishSkill:
         )
         assert resp.json()["skill_id"] == str(new_skill.id)
 
-    @patch("decision_hub.api.registry_routes._build_analyze_prompt_fn", return_value=None)
-    @patch("decision_hub.api.registry_routes._build_analyze_fn", return_value=None)
+    @patch("decision_hub.api.registry_service._build_analyze_prompt_fn", return_value=None)
+    @patch("decision_hub.api.registry_service._build_analyze_fn", return_value=None)
     @patch("decision_hub.api.registry_routes.find_version")
     @patch("decision_hub.api.registry_routes.find_skill")
     @patch("decision_hub.api.registry_routes.compute_checksum")
-    @patch("decision_hub.api.registry_routes.find_org_member")
-    @patch("decision_hub.api.registry_routes.find_org_by_slug")
+    @patch("decision_hub.api.registry_service.find_org_member")
+    @patch("decision_hub.api.registry_service.find_org_by_slug")
     def test_publish_duplicate_version(
         self,
         mock_find_org: MagicMock,
@@ -339,12 +339,12 @@ class TestPublishSkill:
         assert resp.status_code == 409
         assert "already exists" in resp.json()["detail"]
 
-    @patch("decision_hub.api.registry_routes._build_analyze_prompt_fn", return_value=None)
-    @patch("decision_hub.api.registry_routes._build_analyze_fn", return_value=None)
-    @patch("decision_hub.api.registry_routes.insert_audit_log")
-    @patch("decision_hub.api.registry_routes.upload_skill_zip")
-    @patch("decision_hub.api.registry_routes.find_org_member")
-    @patch("decision_hub.api.registry_routes.find_org_by_slug")
+    @patch("decision_hub.api.registry_service._build_analyze_prompt_fn", return_value=None)
+    @patch("decision_hub.api.registry_service._build_analyze_fn", return_value=None)
+    @patch("decision_hub.api.registry_service.insert_audit_log")
+    @patch("decision_hub.api.registry_service.upload_skill_zip")
+    @patch("decision_hub.api.registry_service.find_org_member")
+    @patch("decision_hub.api.registry_service.find_org_by_slug")
     def test_publish_gauntlet_blocks_dangerous_skill(
         self,
         mock_find_org: MagicMock,
@@ -375,12 +375,12 @@ class TestPublishSkill:
         assert resp.status_code == 422
         assert "malformed" in resp.json()["detail"].lower()
 
-    @patch("decision_hub.api.registry_routes._build_analyze_prompt_fn", return_value=None)
-    @patch("decision_hub.api.registry_routes._build_analyze_fn", return_value=None)
-    @patch("decision_hub.api.registry_routes.insert_audit_log")
-    @patch("decision_hub.api.registry_routes.upload_skill_zip")
-    @patch("decision_hub.api.registry_routes.find_org_member")
-    @patch("decision_hub.api.registry_routes.find_org_by_slug")
+    @patch("decision_hub.api.registry_service._build_analyze_prompt_fn", return_value=None)
+    @patch("decision_hub.api.registry_service._build_analyze_fn", return_value=None)
+    @patch("decision_hub.api.registry_service.insert_audit_log")
+    @patch("decision_hub.api.registry_service.upload_skill_zip")
+    @patch("decision_hub.api.registry_service.find_org_member")
+    @patch("decision_hub.api.registry_service.find_org_by_slug")
     def test_publish_gauntlet_blocks_suspicious_code(
         self,
         mock_find_org: MagicMock,
@@ -629,8 +629,8 @@ class TestDeleteSkillVersion:
     @patch("decision_hub.api.registry_routes.delete_skill_zip")
     @patch("decision_hub.api.registry_routes.delete_version")
     @patch("decision_hub.api.registry_routes.find_skill")
-    @patch("decision_hub.api.registry_routes.find_org_member")
-    @patch("decision_hub.api.registry_routes.find_org_by_slug")
+    @patch("decision_hub.api.registry_service.find_org_member")
+    @patch("decision_hub.api.registry_service.find_org_by_slug")
     def test_delete_success(
         self,
         mock_find_org: MagicMock,
@@ -668,8 +668,8 @@ class TestDeleteSkillVersion:
         resp = client.delete("/v1/skills/test-org/my-skill/1.0.0")
         assert resp.status_code == 401
 
-    @patch("decision_hub.api.registry_routes.find_org_member")
-    @patch("decision_hub.api.registry_routes.find_org_by_slug")
+    @patch("decision_hub.api.registry_service.find_org_member")
+    @patch("decision_hub.api.registry_service.find_org_by_slug")
     def test_delete_forbidden_for_member_role(
         self,
         mock_find_org: MagicMock,
@@ -693,8 +693,8 @@ class TestDeleteSkillVersion:
         assert resp.status_code == 403
         assert "owners and admins" in resp.json()["detail"]
 
-    @patch("decision_hub.api.registry_routes.find_org_member")
-    @patch("decision_hub.api.registry_routes.find_org_by_slug")
+    @patch("decision_hub.api.registry_service.find_org_member")
+    @patch("decision_hub.api.registry_service.find_org_by_slug")
     def test_delete_forbidden_for_non_member(
         self,
         mock_find_org: MagicMock,
@@ -715,7 +715,7 @@ class TestDeleteSkillVersion:
 
         assert resp.status_code == 403
 
-    @patch("decision_hub.api.registry_routes.find_org_by_slug")
+    @patch("decision_hub.api.registry_service.find_org_by_slug")
     def test_delete_org_not_found(
         self,
         mock_find_org: MagicMock,
@@ -733,8 +733,8 @@ class TestDeleteSkillVersion:
         assert resp.status_code == 404
 
     @patch("decision_hub.api.registry_routes.find_skill")
-    @patch("decision_hub.api.registry_routes.find_org_member")
-    @patch("decision_hub.api.registry_routes.find_org_by_slug")
+    @patch("decision_hub.api.registry_service.find_org_member")
+    @patch("decision_hub.api.registry_service.find_org_by_slug")
     def test_delete_skill_not_found(
         self,
         mock_find_org: MagicMock,
@@ -760,8 +760,8 @@ class TestDeleteSkillVersion:
 
     @patch("decision_hub.api.registry_routes.delete_version")
     @patch("decision_hub.api.registry_routes.find_skill")
-    @patch("decision_hub.api.registry_routes.find_org_member")
-    @patch("decision_hub.api.registry_routes.find_org_by_slug")
+    @patch("decision_hub.api.registry_service.find_org_member")
+    @patch("decision_hub.api.registry_service.find_org_by_slug")
     def test_delete_version_not_found(
         self,
         mock_find_org: MagicMock,
@@ -791,8 +791,8 @@ class TestDeleteSkillVersion:
     @patch("decision_hub.api.registry_routes.delete_skill_zip")
     @patch("decision_hub.api.registry_routes.delete_version")
     @patch("decision_hub.api.registry_routes.find_skill")
-    @patch("decision_hub.api.registry_routes.find_org_member")
-    @patch("decision_hub.api.registry_routes.find_org_by_slug")
+    @patch("decision_hub.api.registry_service.find_org_member")
+    @patch("decision_hub.api.registry_service.find_org_by_slug")
     def test_delete_allowed_for_admin(
         self,
         mock_find_org: MagicMock,
@@ -1019,8 +1019,8 @@ class TestDeleteAllVersions:
     @patch("decision_hub.api.registry_routes.delete_skill_record")
     @patch("decision_hub.api.registry_routes.delete_all_versions")
     @patch("decision_hub.api.registry_routes.find_skill")
-    @patch("decision_hub.api.registry_routes.find_org_member")
-    @patch("decision_hub.api.registry_routes.find_org_by_slug")
+    @patch("decision_hub.api.registry_service.find_org_member")
+    @patch("decision_hub.api.registry_service.find_org_by_slug")
     def test_delete_all_success(
         self,
         mock_find_org: MagicMock,
@@ -1063,8 +1063,8 @@ class TestDeleteAllVersions:
         resp = client.delete("/v1/skills/test-org/my-skill")
         assert resp.status_code == 401
 
-    @patch("decision_hub.api.registry_routes.find_org_member")
-    @patch("decision_hub.api.registry_routes.find_org_by_slug")
+    @patch("decision_hub.api.registry_service.find_org_member")
+    @patch("decision_hub.api.registry_service.find_org_by_slug")
     def test_delete_all_forbidden_for_member(
         self,
         mock_find_org: MagicMock,
@@ -1087,7 +1087,7 @@ class TestDeleteAllVersions:
 
         assert resp.status_code == 403
 
-    @patch("decision_hub.api.registry_routes.find_org_by_slug")
+    @patch("decision_hub.api.registry_service.find_org_by_slug")
     def test_delete_all_org_not_found(
         self,
         mock_find_org: MagicMock,
@@ -1105,8 +1105,8 @@ class TestDeleteAllVersions:
         assert resp.status_code == 404
 
     @patch("decision_hub.api.registry_routes.find_skill")
-    @patch("decision_hub.api.registry_routes.find_org_member")
-    @patch("decision_hub.api.registry_routes.find_org_by_slug")
+    @patch("decision_hub.api.registry_service.find_org_member")
+    @patch("decision_hub.api.registry_service.find_org_by_slug")
     def test_delete_all_skill_not_found(
         self,
         mock_find_org: MagicMock,
@@ -1134,8 +1134,8 @@ class TestDeleteAllVersions:
     @patch("decision_hub.api.registry_routes.delete_skill_record")
     @patch("decision_hub.api.registry_routes.delete_all_versions")
     @patch("decision_hub.api.registry_routes.find_skill")
-    @patch("decision_hub.api.registry_routes.find_org_member")
-    @patch("decision_hub.api.registry_routes.find_org_by_slug")
+    @patch("decision_hub.api.registry_service.find_org_member")
+    @patch("decision_hub.api.registry_service.find_org_by_slug")
     def test_delete_all_allowed_for_admin(
         self,
         mock_find_org: MagicMock,
