+++ shared/pyproject.toml
@@ -1,0 +1,15 @@
+[project]
+name = "dhub-core"
+version = "0.1.0"
+description = "Shared domain models and manifest parsing for Decision Hub"
+requires-python = ">=3.11"
+dependencies = [
+    "pyyaml>=6.0.0",
+]
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["src/dhub_core"]
